envelope.c: Pre-scale decay gap to eliminate 5% step to sustain.
In "false exponential" envelopes (which emulate the R-C charging of analog ADSR), we had a hard switch from (wherever the curve had got to) to (the sustain level) at the end of the Decay interval.  But because we had only got (1 - exp(-3.0)) of the way through the gap, there was about 5% of the gap remaining, which manifested as a sudden drop in level.  Here, we make the target gap a little bit larger, so that when we've traversed 95% of it at the end of the decay time, we're right at the sustain level.