amy.c: Fix wrap-around clipping at relatively low amplitudes.
There's a final high-pass filter to remove very low, large-amplitude excursions from some voices.  Turns out it was overflowing the +/-16 limit of MUL4 output, but doesn't show problems under reasonable conditions with MUL8 (full +/- 256 range).  

However, since the value is constant and not precise, I was able to replace it with a shift anyway, to avoid any unnecessary quantization loss.